A partial port of the libvpx vp8 encoder to java + webm improvements
- The MKV muxer is now now waiting to write the file to the disk till
the finish() call. Instead it writes after every keyframe. This allows
larger files
- Preexisting vpx code was merged with the new libvpx code
- A new yuv to webm transcoder sample was added
- The libvpx port allows multiple threads to encode in parralel with
multiple encoder objects (but a single one does not do multi-threading)
- Here are the features that were not ported:
-- CPU specific optimisations
-- VP9 encoder/decoder
-- VP8 decoder
-- VP8 encoder postprocessing, spatial-resampling, error-concealment,
multi-resolution-encoding, temporal-denoising and multi threading
- Many constructs for the above omitted features are still partially
ported, they are just not tested and only used to the extent that the
basic features of the encoder needed it.

+++ src/main/java/org/jcodec/codecs/vpx/vp9/CodedSuperBlock.java
@@ -129,7 +129,7 @@ public class CodedSuperBlock {
 //		System.out.print(String.format("PARTITION [%d,%d,%d]", miCol, miRow, blkSize));
         int ctx = calcPartitionContext(miCol, miRow, blkSize, c);
 //		System.out.printf(String.format(", ctx=%d\n", ctx));
-        int[] probs = c.getPartitionProbs()[ctx];
+        short[] probs = c.getPartitionProbs()[ctx];
         int halfBlk = (1 << blkSize) >> 1;
         boolean rightEdge = miCol + halfBlk >= c.getMiTileWidth();
         boolean bottomEdge = miRow + halfBlk >= c.getMiTileHeight();
